combined `builtins.dict` tests with those exclusive to `>=3.9`
